pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astra-agent"
version = "0.1.0"
description = "AI-powered coding agent for repository analysis and modification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "pathspec>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
astra = "astra.cli:main"

[tool.setuptools.packages.find]
include = ["astra*"]
